update: useContext 리펙토링 및 불필요한 코드 삭제
