+++ src/components/PokemonCard.jsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 
-const PokemonCard = ({ onAdd, pokemon, onRemove, isSelected }) => {
+const PokemonCard = ({ onAdd, pokemon }) => {
   const { img_url, korean_name, id } = pokemon;
   const navigate = useNavigate();
 
@@ -13,25 +13,14 @@ const PokemonCard = ({ onAdd, pokemon, onRemove, isSelected }) => {
     >
       <img src={img_url} alt={korean_name} />
       <p>{korean_name}</p>
-      {isSelected ? (
-        <Button
-          onClick={(e) => {
-            e.stopPropagation();
-            onRemove(id);
-          }}
-        >
-          삭제
-        </Button>
-      ) : (
-        <Button
-          onClick={(e) => {
-            e.stopPropagation();
-            onAdd(pokemon);
-          }}
-        >
-          추가
-        </Button>
-      )}
+      <Button
+        onClick={(e) => {
+          e.stopPropagation();
+          onAdd(pokemon);
+        }}
+      >
+        추가
+      </Button>
     </Card>
   );
 };
